Set ROLLUP_INSPECT in HTIF's fromhost data field
Before running a machine to process an InspectState RPC, the
machine has just yielded manual. The rollup machine manager
must set ROLLUP_INSPECT as the data field in HTIF's fromhost CSR
as the acknowledgement from that yield manual so the program running
inside knows the next request is an InspectState, rather than an
AdvanceState.